Beim Install Wizard gibt es nun einen zusätzlichen Step zum aktivieren der Module

+++ modules/install/class_install.php
@@ -181,7 +181,7 @@ class Install {
     global $db, $xml, $func;
 
     // Tabelle Modules leeren um Module zu deinstallieren
-    if($_GET["action"] == "wizard") {
+    if($_GET["action"] == "wizard" and $_GET['step'] < 9) {
       $db->qry("TRUNCATE TABLE %prefix%modules");
     }
     $db->qry("TRUNCATE %prefix%plugin");
@@ -687,6 +687,63 @@ class Install {
     $file = "modules/install/update/update202.php";
     if(file_exists($file)) include($file);
 */
+  }
+
+  function getModConfigLine($row, $showLinks = 1) {
+    global $smarty, $db;
+    
+    $smarty->assign('name', $row['name']);
+    $smarty->assign('caption', $row['caption']);
+    $smarty->assign('description', $row['description']);
+    $smarty->assign('version', $row["version"]);
+
+    if ($row["email"]) $author = "<a href=\"mailto:{$row["email"]}\">{$row["author"]}</a>";
+    else $author = $row["author"];
+    $smarty->assign('author', $author);
+
+    if ($row["reqphp"] or $row["reqmysql"]) {
+      $req = '<br /><u>'. t('Benötigt') .'</u>: ';
+      if ($row["reqphp"]) $req .= 'PHP >= '. $row["reqphp"] .' ';
+      if ($row["reqmysql"]) $req .= 'MySQL >= '. $row["reqmysql"] .' ';
+    } else $req = '';
+    $smarty->assign('req', $req);
+
+    $active = '';
+    if ($row["active"]) $active = ' checked="checked"';
+    $smarty->assign('active', $active);
+
+    $changeable = '';
+    if (!$row["changeable"]) $changeable = ' disabled';
+    $smarty->assign('changeable', $changeable);
+
+    ($row["state"] == "Stable")? $state = $row["state"] : $state = "<font color=\"red\">{$row["state"]}</font>";
+    $smarty->assign('state', $state);
+
+    (file_exists("modules/{$row["name"]}/icon.gif"))? $img = "modules/{$row["name"]}/icon.gif" : $img = "modules/sample/icon.gif";
+    $smarty->assign('img', $img);
+
+    $smarty->assign('showLinks', $showLinks);
+    if ($showLinks) {
+      $find_config = $db->qry_first("SELECT cfg_key FROM %prefix%config WHERE (cfg_module = %string%)", $row["name"]);
+      if ($find_config["cfg_key"] != '') $settings_link = " | <a href=\"index.php?mod=install&action=mod_cfg&step=10&module={$row["name"]}\">". t('Konfig.') ."</a>";
+      else $settings_link = "";
+      $smarty->assign('settings_link', $settings_link);
+
+      $find_mod = $db->qry_first("SELECT module FROM %prefix%menu WHERE module=%string%", $row["name"]);
+      if ($find_mod["module"]) $menu_link = " | <a href=\"index.php?mod=install&action=mod_cfg&step=30&module={$row["name"]}\">". t('Menü') ."</a>";
+      else $menu_link = "";
+      $smarty->assign('menu_link', $menu_link);
+
+      if (file_exists("modules/{$row["name"]}/mod_settings/db.xml")) $db_link = " | <a href=\"index.php?mod=install&action=mod_cfg&step=40&module={$row["name"]}\">". t('DB') ."</a>";
+      else $db_link = "";
+      $smarty->assign('db_link', $db_link);
+
+      if (file_exists("modules/{$row["name"]}/docu/{$language}_help.php")) $help_link = " | <a href=\"#\" onclick=\"javascript:var w=window.open('index.php?mod=helplet&action=helplet&design=base&module={$row["name"]}&helpletid=help','_blank','width=700,height=500,resizable=no,scrollbars=yes');\" class=\"Help\">?</a>";
+      else $help_link = '';
+      $smarty->assign('help_link', $help_link);
+    }
+    
+    return $smarty->fetch('modules/install/templates/module.htm');
   }
 }
 ?>
